move iterator to queue.rs
We're going to make the fields of the Queue private and hence
we won't be able to access them from another module. As the
iterator requires these fields to be public to be initialized,
we can just move the implementation where the definition is.
